tasktracker/userprofile/views.py
from rest_framework import viewsets
from rest_framework import permissions
from .serializers import ProfileSerializer
from .models import Profile
from rest_framework import mixins


class ProfileViewSet(mixins.CreateModelMixin,
                   mixins.RetrieveModelMixin,
                   mixins.UpdateModelMixin,
                   mixins.DestroyModelMixin,
                   mixins.ListModelMixin,
                   viewsets.GenericViewSet):
    queryset = Profile.objects.all()
    serializer_class = ProfileSerializer
    permission_classes = [permissions.IsAuthenticated]
