requirements.txt
asgiref==3.7.2
Django==4.1.10
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
dnspython==2.4.2
environ==1.0
PyJWT==2.8.0
pymongo==3.12.1
pytz==2023.3
sqlparse==0.2.4
